Adding long content lambda extension to bypass step functions size limitation.
The Lambda extension is implemented as a proxy that intercept the
message in and out of the lambda functions. When it detect a long text
in one of the content elements it will store it in a DynamoDB (DDB)
table and replace the value with a reference to the DynamoDB record. On
the other side, when it detects a content value with a reference to a
record, it will retrieve if from DDB and replace the reference with the
original content. This way the call LLM and tool lambda functions don't
need to make any change in their implementation to benefit from the long
content feature.